refactor: moved interface UserFilters
